일급 컬렉션으로 수정
