+++ src/test/java/nextstep/subway/unit/SectionServiceTest.java
@@ -53,7 +53,7 @@ public class SectionServiceTest {
         sectionService.addSection(line.getId(), sectionDto);
 
         // then
-        assertThat(line.getSections()).hasSize(2)
+        assertThat(line.getSections().getSections()).hasSize(2)
                 .extracting("upStation.name", "downStation.name")
                 .containsExactlyInAnyOrder(
                         Tuple.tuple("당고개역", "사당역"),
@@ -71,7 +71,7 @@ public class SectionServiceTest {
         sectionService.addSection(line.getId(), sectionDto);
 
         // then
-        assertThat(line.getSections()).hasSize(2)
+        assertThat(line.getSections().getSections()).hasSize(2)
                 .extracting("upStation.name", "downStation.name")
                 .containsExactlyInAnyOrder(
                         Tuple.tuple("사당역", "당고개역"),
@@ -89,7 +89,7 @@ public class SectionServiceTest {
         sectionService.addSection(line.getId(), sectionDto);
 
         // then
-        assertThat(line.getSections()).hasSize(2)
+        assertThat(line.getSections().getSections()).hasSize(2)
                 .extracting("upStation.name", "downStation.name")
                 .containsExactlyInAnyOrder(
                         Tuple.tuple("당고개역", "이수역"),
@@ -161,7 +161,7 @@ public class SectionServiceTest {
         sectionService.removeSection(line.getId(), 사당역.getId());
 
         // then : 결과 확인
-        assertThat(line.getSections()).hasSize(1)
+        assertThat(line.getSections().getSections()).hasSize(1)
                 .extracting("upStation.name", "downStation.name")
                 .containsExactlyInAnyOrder(
                         Tuple.tuple("사당역", "이수역")
